주문원장 상품 원장 분리 마이그레이션

+++ order/src/main/java/com/example/order/domain/Product.java
@@ -36,7 +36,7 @@ public class Product {
                                                 long sizeId,
                                                 int amount,
                                                 int orderAmount,
-                                                int couponId,
+                                                Long couponId,
                                                 int salePrice,
                                                 LocalDateTime createAt,
                                                 LocalDateTime updateAt
